dashboard: extract ChangeRoleModal and InviteMemberModal from MembersPage
